Seperate routes in books.js file
